merged: changes from linuxport revs r22366-r23076
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing@23150 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/guilib/common/SDLJoystick.h
@@ -7,6 +7,9 @@
 
 #ifdef HAS_SDL_JOYSTICK
 
+#include <SDL/SDL_joystick.h>
+#include <SDL/SDL_events.h>
+
 #define MAX_AXES 64
 
 #define JACTIVE_BUTTON 0x00000001
@@ -63,7 +66,7 @@ private:
   int m_SafeRange; // dead zone
   Uint32 m_pressTicksButton;
   Uint32 m_pressTicksHat;
-  WORD m_ActiveFlags;
+  Uint8 m_ActiveFlags;
   std::vector<SDL_Joystick*> m_Joysticks;
   std::vector<std::string> m_JoystickNames;
 };
